Issue 523:grpc changes for events (#569)
* made changes for etcd k8 deployment

* addressed the review comments

* addressed the review comments

* added scripts for generating and using certificates and keys for etcd

* added functions for implementing the gRPC server and client

* added cli data collection and changed the service names

* refactored code

* refactored implemetation

* corrected the cli issue

* refactored common code

* made init a global function

* moved CollectCLIArgs to config package

* updated create_helm_packages.sh for etcd

* corrected the roles group all for etcd

* corrected the certificate generation

* corrected the certificate generation

* added code for adding and deleting etcd conf and data folders

* corrected the stateful.yaml files

* replaced proto files for go micro with grpc diles

* corrected the svc-managers gRPC server

* corrected the implementation

* corrected the unit test case

* corrected the unit test case

* corrected the client calls to svc-managers

* corrected the deployment

* commented the cli call

* added client service implementation

* added logs for debug

* debugging

* changed the port

* added log

* corrected the tls for etcd

* removed unwanted logs

* Modification in systems for grpc

* Committing after running go tools

* committing after running go tools

* removed unnecessary logs

* corrected variable in systems

* Committing proto files

* corrected start_system command

* Delete systems.micro.go

* Delete chassis.micro.go

* Added framework in command

* Added framework in command

* committing after executing go tools

* grpc changes for events

* grpc changes for events

* Committing after running go tools

* Committing after running go tools

* Modification for grpc in svc-aggregation

* Modification for grpc in svc-aggregation

* modificationin commn for event subscription

* grpc changes for task

* Committing after running go tools

* added etcd in registy

* Committing after running go tools

* modifed registry with etcd

* modification afterexecuting go tools

* executed go fmt

* correct log package

+++ svc-api/rpc/task.go
@@ -29,7 +29,12 @@ import (
 
 // DeleteTaskRequest will do the rpc calls for the svc-task DeleteTask
 func DeleteTaskRequest(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	asService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	asService := taskproto.NewGetTaskServiceClient(conn)
 	// Call the DeleteTask
 	rsp, err := asService.DeleteTask(context.TODO(), req)
 	if err != nil {
@@ -47,7 +52,12 @@ func DeleteTaskRequest(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, 
 
 // GetTaskRequest will do the rpc calls for the svc-task GetTaskStatus
 func GetTaskRequest(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	asService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	asService := taskproto.NewGetTaskServiceClient(conn)
 	// Call the GetTasks
 	rsp, err := asService.GetTasks(context.TODO(), req)
 	if err != nil {
@@ -65,7 +75,12 @@ func GetTaskRequest(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, err
 
 // GetSubTasks will do the rpc calls for the svc-task GetSubTasks
 func GetSubTasks(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	tService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	tService := taskproto.NewGetTaskServiceClient(conn)
 	// Call the GetSubTasks
 	rsp, err := tService.GetSubTasks(context.TODO(), req)
 	if err != nil {
@@ -83,7 +98,12 @@ func GetSubTasks(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error)
 
 // GetSubTask will do the rpc calls for the svc-task GetSubTask
 func GetSubTask(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	tService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	tService := taskproto.NewGetTaskServiceClient(conn)
 	// Call the GetSubTask
 	rsp, err := tService.GetSubTask(context.TODO(), req)
 	if err != nil {
@@ -101,7 +121,12 @@ func GetSubTask(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) 
 
 // GetTaskMonitor will do the rpc calls for the svc-task GetTaskMonitor
 func GetTaskMonitor(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	tService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	tService := taskproto.NewGetTaskServiceClient(conn)
 	// perform rpc call to svc-task to get TaskMonitor resource
 	rsp, err := tService.GetTaskMonitor(context.TODO(), req)
 	if err != nil {
@@ -119,7 +144,12 @@ func GetTaskMonitor(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, err
 
 //TaskCollection will perform the rpc call to svc-task TaskCollection
 func TaskCollection(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	tService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	tService := taskproto.NewGetTaskServiceClient(conn)
 	// perform rpc call to svc-task to get TaskCollection resource
 	rsp, err := tService.TaskCollection(context.TODO(), req)
 	if err != nil {
@@ -137,7 +167,12 @@ func TaskCollection(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, err
 
 //GetTaskService will perform the rpc call to svc-task GetTaskService
 func GetTaskService(req *taskproto.GetTaskRequest) (*taskproto.TaskResponse, error) {
-	tService := taskproto.NewGetTaskService(services.Tasks, services.Service.Client())
+	conn, connErr := services.ODIMService.Client(services.Tasks)
+	if connErr != nil {
+		return nil, fmt.Errorf("Failed to create client connection: %v", connErr)
+	}
+	defer conn.Close()
+	tService := taskproto.NewGetTaskServiceClient(conn)
 	// perform rpc call to svc-task to get TaskService resource
 	rsp, err := tService.GetTaskService(context.TODO(), req)
 	if err != nil {
